+++ temba/orgs/tests.py
@@ -48,7 +48,6 @@ from temba.tickets.models import TicketExport
 from temba.triggers.models import Trigger
 from temba.utils import json, languages
 from temba.utils.uuid import uuid4
-from temba.utils.views.mixins import TEMBA_MENU_SELECTION
 
 from .context_processors import RolePermsWrapper
 from .models import (
@@ -136,8 +135,10 @@ class InvitationTest(TembaTest):
         self.assertEqual("RapidPro Invitation", mail.outbox[0].subject)
         self.assertIn(f"https://app.rapidpro.io/org/join/{invitation.secret}/", mail.outbox[0].body)
 
-        invitation.release()
+        user = User.create("invitededitor@nyaruka.com", "Bob", "", "Qwerty123", "en-US")
+        invitation.accept(user)
 
+        self.assertEqual(1, self.admin.notifications.count())
         self.assertFalse(invitation.is_active)
 
     def test_expire_task(self):
@@ -631,44 +632,8 @@ class UserTest(TembaTest):
         response = self.client.post(confirm_url, {"password": "Qwerty123"})
         self.assertRedirect(response, "/msg/")
 
-    def test_ui_permissions(self):
-        # non-logged in users can't go here
-        response = self.client.get(reverse("orgs.user_list"))
-        self.assertRedirect(response, "/users/login/")
-        response = self.client.post(reverse("orgs.user_delete", args=(self.editor.pk,)), dict(delete=True))
-        self.assertRedirect(response, "/users/login/")
-
-        # either can admins
-        self.login(self.admin)
-        response = self.client.get(reverse("orgs.user_list"))
-        self.assertRedirect(response, "/users/login/")
-        response = self.client.post(reverse("orgs.user_delete", args=(self.editor.pk,)), dict(delete=True))
-        self.assertRedirect(response, "/users/login/")
-
-        self.editor.refresh_from_db()
-        self.assertTrue(self.editor.is_active)
-
-    def test_ui_management(self):
-        # only customer support gets in on this sweet action
-        self.login(self.customer_support)
-
-        # one of our users should belong to a bunch of orgs
-        for i in range(5):
-            org = Org.objects.create(
-                name=f"Org {i}", timezone=ZoneInfo("Africa/Kigali"), created_by=self.user, modified_by=self.user
-            )
-            org.add_user(self.admin, OrgRole.ADMINISTRATOR)
-
-        response = self.client.get(reverse("orgs.user_list"))
-        self.assertEqual(200, response.status_code)
-
-        response = self.client.post(reverse("orgs.user_delete", args=(self.editor.pk,)), {"delete": True})
-        self.assertEqual(reverse("orgs.user_list"), response["Temba-Success"])
-
-        self.editor.refresh_from_db()
-        self.assertFalse(self.editor.is_active)
-
-    def test_release(self):
+    @mock_mailroom
+    def test_release(self, mr_mocks):
         token = APIToken.create(self.org, self.admin)
 
         # admin doesn't "own" any orgs
@@ -1856,34 +1821,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         menu = self.client.get(menu_url).json()["results"]
         self.assertEqual("tomato", menu[8]["bubble"])
 
-    def test_read(self):
-        read_url = reverse("orgs.org_read", args=[self.org.id])
-
-        # make our second org a child
-        self.org2.parent = self.org
-        self.org2.save()
-
-        response = self.assertStaffOnly(read_url)
-
-        # we should have a child in our context
-        self.assertEqual(1, len(response.context["children"]))
-
-        # we should have options to flag and suspend
-        self.assertContentMenu(read_url, self.customer_support, ["Edit", "Flag", "Suspend", "Verify", "-", "Service"])
-
-        # flag and content menu option should be inverted
-        self.org.flag()
-        self.org.suspend()
-
-        self.assertContentMenu(
-            read_url, self.customer_support, ["Edit", "Unflag", "Unsuspend", "Verify", "-", "Service"]
-        )
-
-        # no menu for inactive orgs
-        self.org.is_active = False
-        self.org.save()
-        self.assertContentMenu(read_url, self.customer_support, [])
-
     def test_workspace(self):
         workspace_url = reverse("orgs.org_workspace")
 
@@ -2166,6 +2103,9 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         invitation.refresh_from_db()
         self.assertFalse(invitation.is_active)
 
+        self.assertEqual(1, self.admin.notifications.filter(notification_type="invitation:accepted").count())
+        self.assertEqual(2, self.org.get_users(roles=[OrgRole.EDITOR]).count())
+
     def test_join_accept(self):
         # only authenticated users can access page
         response = self.client.get(reverse("orgs.org_join_accept", args=["invalid"]))
@@ -2207,6 +2147,9 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
         invitation.refresh_from_db()
         self.assertFalse(invitation.is_active)
+
+        self.assertEqual(1, self.admin.notifications.filter(notification_type="invitation:accepted").count())
+        self.assertEqual(2, self.org.get_users(roles=[OrgRole.EDITOR]).count())
 
     def test_org_grant(self):
         grant_url = reverse("orgs.org_grant")
@@ -2860,7 +2803,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertRedirect(self.requestView(start_url, self.agent), "/ticket/")
 
         # now try as customer support
-        self.assertRedirect(self.requestView(start_url, self.customer_support), "/org/manage/")
+        self.assertRedirect(self.requestView(start_url, self.customer_support), "/staff/org/")
 
         # if org isn't set, we redirect instead to choose view
         self.client.logout()
@@ -2895,7 +2838,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         self.assertContains(response, "No workspaces for this account, please contact your administrator.")
 
         # unless they are staff
-        self.assertRedirect(self.requestView(choose_url, self.customer_support), "/org/manage/")
+        self.assertRedirect(self.requestView(choose_url, self.customer_support), "/staff/org/")
 
         # turn editor into a multi-org user
         self.org2.add_user(self.editor, OrgRole.EDITOR)
@@ -2986,107 +2929,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
         child.refresh_from_db()
         self.assertFalse(child.is_active)
 
-    def test_administration(self):
-        self.setUpLocations()
-
-        manage_url = reverse("orgs.org_manage")
-        update_url = reverse("orgs.org_update", args=[self.org.id])
-
-        self.assertStaffOnly(manage_url)
-        self.assertStaffOnly(update_url)
-
-        def assertOrgFilter(query: str, expected_orgs: list):
-            response = self.client.get(manage_url + query)
-            self.assertIsNotNone(response.headers.get(TEMBA_MENU_SELECTION, None))
-            self.assertEqual(expected_orgs, list(response.context["object_list"]))
-
-        assertOrgFilter("", [self.org2, self.org])
-        assertOrgFilter("?filter=all", [self.org2, self.org])
-        assertOrgFilter("?filter=xxxx", [self.org2, self.org])
-        assertOrgFilter("?filter=flagged", [])
-        assertOrgFilter("?filter=anon", [])
-        assertOrgFilter("?filter=suspended", [])
-        assertOrgFilter("?filter=verified", [])
-
-        self.org.flag()
-
-        assertOrgFilter("?filter=flagged", [self.org])
-
-        self.org2.verify()
-
-        assertOrgFilter("?filter=verified", [self.org2])
-
-        # and can go to our org
-        response = self.client.get(update_url)
-        self.assertEqual(200, response.status_code)
-        self.assertEqual(
-            [
-                "name",
-                "features",
-                "is_anon",
-                "channels_limit",
-                "fields_limit",
-                "globals_limit",
-                "groups_limit",
-                "labels_limit",
-                "teams_limit",
-                "topics_limit",
-                "loc",
-            ],
-            list(response.context["form"].fields.keys()),
-        )
-
-        # make some changes to our org
-        response = self.client.post(
-            update_url,
-            {
-                "name": "Temba II",
-                "features": ["new_orgs"],
-                "is_anon": False,
-                "channels_limit": 20,
-                "fields_limit": 300,
-                "globals_limit": "",
-                "groups_limit": 400,
-                "labels_limit": "",
-                "teams_limit": "",
-                "topics_limit": "",
-            },
-        )
-        self.assertEqual(302, response.status_code)
-
-        self.org.refresh_from_db()
-        self.assertEqual("Temba II", self.org.name)
-        self.assertEqual(["new_orgs"], self.org.features)
-        self.assertEqual(self.org.get_limit(Org.LIMIT_FIELDS), 300)
-        self.assertEqual(self.org.get_limit(Org.LIMIT_GLOBALS), 250)  # uses default
-        self.assertEqual(self.org.get_limit(Org.LIMIT_GROUPS), 400)
-        self.assertEqual(self.org.get_limit(Org.LIMIT_CHANNELS), 20)
-
-        # flag org
-        self.client.post(update_url, {"action": "flag"})
-        self.org.refresh_from_db()
-        self.assertTrue(self.org.is_flagged)
-
-        # unflag org
-        self.client.post(update_url, {"action": "unflag"})
-        self.org.refresh_from_db()
-        self.assertFalse(self.org.is_flagged)
-
-        # suspend org
-        self.client.post(update_url, {"action": "suspend"})
-        self.org.refresh_from_db()
-        self.assertTrue(self.org.is_suspended)
-
-        # unsuspend org
-        self.client.post(update_url, {"action": "unsuspend"})
-        self.org.refresh_from_db()
-        self.assertFalse(self.org.is_suspended)
-
-        # verify
-        self.client.post(update_url, {"action": "verify"})
-        self.org.refresh_from_db()
-        self.assertTrue(self.org.is_verified)
-
     def test_urn_schemes(self):
         # remove existing channels
         Channel.objects.all().update(is_active=False, org=None)
@@ -3164,7 +3006,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
         # posting invalid org just redirects back to manage page
         response = self.client.post(service_url, {"other_org": 325253256})
-        self.assertRedirect(response, "/org/manage/")
+        self.assertRedirect(response, "/staff/org/")
 
         # then service our org
         response = self.client.get(service_url, {"other_org": self.org.id})
@@ -3198,7 +3040,7 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
         # stop servicing
         response = self.client.post(service_url, {})
-        self.assertRedirect(response, "/org/manage/")
+        self.assertRedirect(response, "/staff/org/")
         self.assertIsNone(self.client.session["org_id"])
         self.assertFalse(self.client.session["servicing"])
 
@@ -3292,123 +3134,6 @@ class OrgCRUDLTest(TembaTest, CRUDLTestMixin):
 
 
 class UserCRUDLTest(TembaTest, CRUDLTestMixin):
-    def test_list(self):
-        list_url = reverse("orgs.user_list")
-
-        self.assertStaffOnly(list_url)
-
-        response = self.requestView(list_url, self.customer_support)
-        self.assertEqual(8, len(response.context["object_list"]))
-        self.assertEqual("/staff/users/all", response.headers[TEMBA_MENU_SELECTION])
-
-        response = self.requestView(list_url + "?filter=beta", self.customer_support)
-        self.assertEqual(set(), set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?filter=staff", self.customer_support)
-        self.assertEqual({self.customer_support, self.superuser}, set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?search=admin@nyaruka.com", self.customer_support)
-        self.assertEqual({self.admin}, set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?search=admin@nyaruka.com", self.customer_support)
-        self.assertEqual({self.admin}, set(response.context["object_list"]))
-
-        response = self.requestView(list_url + "?search=Andy", self.customer_support)
-        self.assertEqual({self.admin}, set(response.context["object_list"]))
-
-    def test_read(self):
-        read_url = reverse("orgs.user_read", args=[self.editor.id])
-
-        # this is a customer support only view
-        self.assertStaffOnly(read_url)
-
-        response = self.requestView(read_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-
-    def test_update(self):
-        update_url = reverse("orgs.user_update", args=[self.editor.id])
-
-        # this is a customer support only view
-        self.assertStaffOnly(update_url)
-
-        response = self.requestView(update_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-
-        alphas = Group.objects.get(name="Alpha")
-        betas = Group.objects.get(name="Beta")
-        current_password = self.editor.password
-
-        # submit without new password
-        response = self.requestView(
-            update_url,
-            self.customer_support,
-            post_data={
-                "email": "eddy@nyaruka.com",
-                "first_name": "Edward",
-                "last_name": "",
-                "groups": [alphas.id, betas.id],
-            },
-        )
-        self.assertEqual(302, response.status_code)
-
-        self.editor.refresh_from_db()
-        self.assertEqual("eddy@nyaruka.com", self.editor.email)
-        self.assertEqual("eddy@nyaruka.com", self.editor.username)  # should match email
-        self.assertEqual(current_password, self.editor.password)
-        self.assertEqual("Edward", self.editor.first_name)
-        self.assertEqual("", self.editor.last_name)
-        self.assertEqual({alphas, betas}, set(self.editor.groups.all()))
-
-        # submit with new password and one less group
-        response = self.requestView(
-            update_url,
-            self.customer_support,
-            post_data={
-                "email": "eddy@nyaruka.com",
-                "new_password": "Asdf1234",
-                "first_name": "Edward",
-                "last_name": "",
-                "groups": [alphas.id],
-            },
-        )
-        self.assertEqual(302, response.status_code)
-
-        self.editor.refresh_from_db()
-        self.assertEqual("eddy@nyaruka.com", self.editor.email)
-        self.assertEqual("eddy@nyaruka.com", self.editor.username)
-        self.assertNotEqual(current_password, self.editor.password)
-        self.assertEqual("Edward", self.editor.first_name)
-        self.assertEqual("", self.editor.last_name)
-        self.assertEqual({alphas}, set(self.editor.groups.all()))
-
-    def test_delete(self):
-        delete_url = reverse("orgs.user_delete", args=[self.editor.id])
-
-        # this is a customer support only view
-        self.assertStaffOnly(delete_url)
-
-        response = self.requestView(delete_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-        self.assertNotContains(response, "Nyaruka")  # editor doesn't own this org
-
-        # make editor the owner of the org
-        OrgMembership.objects.filter(org=self.org, role_code=OrgRole.ADMINISTRATOR.code).delete()
-        OrgMembership.objects.filter(org=self.org, role_code=OrgRole.VIEWER.code).delete()
-        OrgMembership.objects.filter(org=self.org, role_code=OrgRole.AGENT.code).delete()
-
-        response = self.requestView(delete_url, self.customer_support)
-        self.assertEqual(200, response.status_code)
-        self.assertContains(response, "Nyaruka")
-
-        response = self.requestView(delete_url, self.customer_support, post_data={})
-        self.assertEqual(reverse("orgs.user_list"), response["Temba-Success"])
-
-        self.editor.refresh_from_db()
-        self.assertFalse(self.editor.is_active)
-
-        self.org.refresh_from_db()
-        self.assertFalse(self.org.is_active)
-
     def test_account(self):
         self.login(self.agent)