+++ temba/orgs/views/views.py
@@ -6,8 +6,8 @@ import iso8601
 import pyotp
 from django_redis import get_redis_connection
 from packaging.version import Version
-from smartmin.users.models import FailedLogin, PasswordHistory, RecoveryToken
-from smartmin.users.views import Login, UserUpdateForm
+from smartmin.users.models import FailedLogin, RecoveryToken
+from smartmin.users.views import Login
 from smartmin.views import (
     SmartCreateView,
     SmartCRUDL,
@@ -23,20 +23,18 @@ from django import forms
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout, update_session_auth_hash
-from django.contrib.auth.models import Group
 from django.contrib.auth.password_validation import validate_password
 from django.contrib.auth.views import LoginView as AuthLoginView
 from django.core.exceptions import ValidationError
 from django.db.models.functions import Lower
 from django.forms import ModelChoiceField
-from django.http import Http404, HttpResponse, HttpResponseRedirect, JsonResponse
+from django.http import Http404, HttpResponseRedirect, JsonResponse
 from django.shortcuts import get_object_or_404, resolve_url
 from django.urls import reverse, reverse_lazy
 from django.utils import timezone
 from django.utils.encoding import DjangoUnicodeDecodeError, force_str
 from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
-from django.views.decorators.csrf import csrf_exempt
 
 from temba.api.models import APIToken, Resthook
 from temba.campaigns.models import Campaign
@@ -44,7 +42,7 @@ from temba.flows.models import Flow
 from temba.formax import FormaxMixin
 from temba.notifications.mixins import NotificationTargetMixin
 from temba.orgs.tasks import send_user_verification_email
-from temba.utils import analytics, get_anonymous_user, json, languages, on_transaction_commit, str_to_bool
+from temba.utils import analytics, json, languages, on_transaction_commit, str_to_bool
 from temba.utils.email import EmailSender, parse_smtp_url
 from temba.utils.fields import (
     ArbitraryJsonChoiceField,
@@ -340,11 +338,7 @@ class InferUserMixin:
 class UserCRUDL(SmartCRUDL):
     model = User
     actions = (
-        "list",
-        "update",
         "edit",
-        "delete",
-        "read",
         "forget",
         "recover",
         "two_factor_enable",
@@ -355,114 +349,6 @@ class UserCRUDL(SmartCRUDL):
         "verify_email",
         "send_verification_email",
     )
-
-    class Read(StaffOnlyMixin, ContextMenuMixin, SpaMixin, SmartReadView):
-        fields = ("email", "date_joined")
-        menu_path = "/staff/users/all"
-
-        def build_context_menu(self, menu):
-            obj = self.get_object()
-            menu.add_modax(
-                _("Edit"),
-                "user-update",
-                reverse("orgs.user_update", args=[obj.id]),
-                title=_("Edit User"),
-                as_button=True,
-            )
-
-            menu.add_modax(
-                _("Delete"),
-                "user-delete",
-                reverse("orgs.user_delete", args=[obj.id]),
-                title=_("Delete User"),
-            )
-
-    class List(StaffOnlyMixin, SpaMixin, SmartListView):
-        fields = ("email", "name", "date_joined")
-        ordering = ("-date_joined",)
-        search_fields = ("email__icontains", "first_name__icontains", "last_name__icontains")
-        filters = (("all", _("All")), ("beta", _("Beta")), ("staff", _("Staff")))
-
-        def derive_menu_path(self):
-            return f"/staff/users/{self.request.GET.get('filter', 'all')}"
-
-        @csrf_exempt
-        def dispatch(self, *args, **kwargs):
-            return super().dispatch(*args, **kwargs)
-
-        def derive_queryset(self, **kwargs):
-            qs = super().derive_queryset(**kwargs).filter(is_active=True).exclude(id=get_anonymous_user().id)
-            obj_filter = self.request.GET.get("filter")
-            if obj_filter == "beta":
-                qs = qs.filter(groups__name="Beta")
-            elif obj_filter == "staff":
-                qs = qs.filter(is_staff=True)
-            return qs
-
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-            context["filter"] = self.request.GET.get("filter", "all")
-            context["filters"] = self.filters
-            return context
-
-    class Update(StaffOnlyMixin, ModalFormMixin, ComponentFormMixin, ContextMenuMixin, SmartUpdateView):
-        class Form(UserUpdateForm):
-            groups = forms.ModelMultipleChoiceField(
-                widget=SelectMultipleWidget(
-                    attrs={"placeholder": _("Optional: Select permissions groups."), "searchable": True}
-                ),
-                queryset=Group.objects.all(),
-                required=False,
-            )
-
-            class Meta:
-                model = User
-                fields = ("email", "new_password", "first_name", "last_name", "groups")
-                help_texts = {"new_password": _("You can reset the user's password by entering a new password here")}
-
-        form_class = Form
-        success_message = "User updated successfully."
-        title = "Update User"
-
-        def pre_save(self, obj):
-            obj.username = obj.email
-            return obj
-
-        def post_save(self, obj):
-            """
-            Make sure our groups are up-to-date
-            """
-            if "groups" in self.form.cleaned_data:
-                obj.groups.clear()
-                for group in self.form.cleaned_data["groups"]:
-                    obj.groups.add(group)
-
-            # if a new password was set, reset our failed logins
-            if "new_password" in self.form.cleaned_data and self.form.cleaned_data["new_password"]:
-                FailedLogin.objects.filter(username__iexact=self.object.username).delete()
-                PasswordHistory.objects.create(user=obj, password=obj.password)
-
-            return obj
-
-    class Delete(StaffOnlyMixin, ModalFormMixin, SmartDeleteView):
-        fields = ("id",)
-        permission = "orgs.user_update"
-        submit_button_name = _("Delete")
-        cancel_url = "@orgs.user_list"
-
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-            context["owned_orgs"] = self.get_object().get_owned_orgs()
-            return context
-
-        def post(self, request, *args, **kwargs):
-            user = self.get_object()
-            user.release(self.request.user)
-
-            messages.info(request, self.derive_success_message())
-            response = HttpResponse()
-            response["Temba-Success"] = reverse("orgs.user_list")
-            return response
 
     class Forget(SmartFormView):
         class Form(forms.Form):
@@ -959,10 +845,10 @@ class InvitationMixin:
 
 
 class OrgCRUDL(SmartCRUDL):
+    model = Org
     actions = (
         "signup",
         "start",
-        "read",
         "edit",
         "edit_sub_org",
         "join",
@@ -973,9 +859,7 @@ class OrgCRUDL(SmartCRUDL):
         "delete_child",
         "manage_accounts",
         "manage_accounts_sub_org",
-        "manage",
         "menu",
-        "update",
         "country",
         "languages",
         "sub_orgs",
@@ -987,8 +871,6 @@ class OrgCRUDL(SmartCRUDL):
         "flow_smtp",
         "workspace",
     )
-
-    model = Org
 
     class Menu(BaseMenuView):
         @classmethod
@@ -1140,13 +1022,13 @@ class OrgCRUDL(SmartCRUDL):
                         menu_id="workspaces",
                         name=_("Workspaces"),
                         icon="workspace",
-                        href=reverse("orgs.org_manage"),
+                        href=reverse("staff.org_list"),
                     ),
                     self.create_menu_item(
                         menu_id="users",
                         name=_("Users"),
                         icon="users",
-                        href=reverse("orgs.user_list"),
+                        href=reverse("staff.user_list"),
                     ),
                 ]
 
@@ -1444,209 +1326,6 @@ class OrgCRUDL(SmartCRUDL):
             context["from_email_custom"] = from_email_custom
             return context
 
-    class Read(StaffOnlyMixin, SpaMixin, ContextMenuMixin, SmartReadView):
-        def build_context_menu(self, menu):
-            obj = self.get_object()
-            if not obj.is_active:
-                return
-
-            menu.add_modax(
-                _("Edit"),
-                "update-workspace",
-                reverse("orgs.org_update", args=[obj.id]),
-                title=_("Edit Workspace"),
-                as_button=True,
-                on_submit="handleWorkspaceUpdated()",
-            )
-
-            if not obj.is_flagged:
-                menu.add_url_post(_("Flag"), f"{reverse('orgs.org_update', args=[obj.id])}?action=flag")
-            else:
-                menu.add_url_post(_("Unflag"), f"{reverse('orgs.org_update', args=[obj.id])}?action=unflag")
-
-            if not obj.is_child:
-                if not obj.is_suspended:
-                    menu.add_url_post(_("Suspend"), f"{reverse('orgs.org_update', args=[obj.id])}?action=suspend")
-                else:
-                    menu.add_url_post(_("Unsuspend"), f"{reverse('orgs.org_update', args=[obj.id])}?action=unsuspend")
-
-            if not obj.is_verified:
-                menu.add_url_post(_("Verify"), f"{reverse('orgs.org_update', args=[obj.id])}?action=verify")
-
-            menu.new_group()
-            menu.add_url_post(
-                _("Service"),
-                f'{reverse("orgs.org_service")}?other_org={obj.id}&next={reverse("msgs.msg_inbox", args=[])}',
-            )
-
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-
-            org = self.get_object()
-
-            users_roles = []
-            for role in OrgRole:
-                role_users = list(org.get_users(roles=[role]).values("id", "email"))
-                if role_users:
-                    users_roles.append(dict(role_display=role.display_plural, users=role_users))
-
-            context["users_roles"] = users_roles
-            context["children"] = Org.objects.filter(parent=org, is_active=True).order_by("-created_on", "name")
-            return context
-
-    class Manage(StaffOnlyMixin, SpaMixin, SmartListView):
-        fields = ("name", "owner", "timezone", "created_on")
-        default_order = ("-created_on",)
-        search_fields = ("name__icontains", "created_by__email__iexact", "config__icontains")
-        link_fields = ("name", "owner")
-        filters = (
-            ("all", _("All"), dict(), ("-created_on",)),
-            ("anon", _("Anonymous"), dict(is_anon=True, is_suspended=False), None),
-            ("flagged", _("Flagged"), dict(is_flagged=True, is_suspended=False), None),
-            ("suspended", _("Suspended"), dict(is_suspended=True), None),
-            ("verified", _("Verified"), dict(config__verified=True, is_suspended=False), None),
-        )
-
-        @csrf_exempt
-        def dispatch(self, *args, **kwargs):
-            return super().dispatch(*args, **kwargs)
-
-        def get_filter(self):
-            obj_filter = self.request.GET.get("filter", "all")
-            for filter in self.filters:
-                if filter[0] == obj_filter:
-                    return filter
-
-        def derive_title(self):
-            filter = self.get_filter()
-            if filter:
-                return filter[1]
-            return super().derive_title()
-
-        def derive_menu_path(self):
-            return f"/staff/{self.request.GET.get('filter', 'all')}"
-
-        def get_owner(self, obj):
-            owner = obj.get_owner()
-            return f"{owner.name} ({owner.email})"
-
-        def derive_queryset(self, **kwargs):
-            qs = super().derive_queryset(**kwargs).filter(is_active=True)
-            filter = self.get_filter()
-            if filter:
-                _, _, filter_kwargs, ordering = filter
-                qs = qs.filter(**filter_kwargs)
-                if ordering:
-                    qs = qs.order_by(*ordering)
-                else:
-                    qs = qs.order_by(*self.default_order)
-            else:
-                qs = qs.filter(is_suspended=False).order_by(*self.default_order)
-
-            return qs
-
-        def derive_ordering(self):
-            # we do this in derive queryset for simplicity
-            return None
-
-        def get_context_data(self, **kwargs):
-            context = super().get_context_data(**kwargs)
-            context["filter"] = self.request.GET.get("filter", "all")
-            context["filters"] = self.filters
-            return context
-
-        def lookup_field_link(self, context, field, obj):
-            if field == "owner":
-                owner = obj.get_owner()
-                return reverse("orgs.user_update", args=[owner.pk])
-            return super().lookup_field_link(context, field, obj)
-
-    class Update(StaffOnlyMixin, ModalFormMixin, ComponentFormMixin, SmartUpdateView):
-        ACTION_FLAG = "flag"
-        ACTION_UNFLAG = "unflag"
-        ACTION_SUSPEND = "suspend"
-        ACTION_UNSUSPEND = "unsuspend"
-        ACTION_VERIFY = "verify"
-
-        class Form(forms.ModelForm):
-            features = forms.MultipleChoiceField(
-                choices=Org.FEATURES_CHOICES, widget=SelectMultipleWidget(), required=False
-            )
-
-            def __init__(self, org, *args, **kwargs):
-                super().__init__(*args, **kwargs)
-
-                self.limits_rows = []
-                self.add_limits_fields(org)
-
-            def clean(self):
-                super().clean()
-
-                limits = dict()
-                for row in self.limits_rows:
-                    if self.cleaned_data.get(row["limit_field_key"]):
-                        limits[row["limit_type"]] = self.cleaned_data.get(row["limit_field_key"])
-
-                self.cleaned_data["limits"] = limits
-
-                return self.cleaned_data
-
-            def add_limits_fields(self, org: Org):
-                for limit_type in settings.ORG_LIMIT_DEFAULTS.keys():
-                    field = forms.IntegerField(
-                        label=limit_type.capitalize(),
-                        required=False,
-                        initial=org.limits.get(limit_type),
-                        widget=forms.TextInput(attrs={"placeholder": _("Limit")}),
-                    )
-                    field_key = f"{limit_type}_limit"
-
-                    self.fields.update(OrderedDict([(field_key, field)]))
-                    self.limits_rows.append({"limit_type": limit_type, "limit_field_key": field_key})
-
-            class Meta:
-                model = Org
-                fields = ("name", "features", "is_anon")
-
-        form_class = Form
-        success_url = "hide"
-
-        def derive_title(self):
-            return None
-
-        def get_form_kwargs(self):
-            kwargs = super().get_form_kwargs()
-            kwargs["org"] = self.get_object()
-            return kwargs
-
-        def post(self, request, *args, **kwargs):
-            if "action" in request.POST:
-                action = request.POST["action"]
-                obj = self.get_object()
-
-                if action == self.ACTION_FLAG:
-                    obj.flag()
-                elif action == self.ACTION_UNFLAG:
-                    obj.unflag()
-                elif action == self.ACTION_SUSPEND:
-                    obj.suspend()
-                elif action == self.ACTION_UNSUSPEND:
-                    obj.unsuspend()
-                elif action == self.ACTION_VERIFY:
-                    obj.verify()
-
-                return HttpResponseRedirect(reverse("orgs.org_read", args=[obj.id]))
-
-            return super().post(request, *args, **kwargs)
-
-        def pre_save(self, obj):
-            obj = super().pre_save(obj)
-
-            cleaned_data = self.form.cleaned_data
-
-            obj.limits = cleaned_data["limits"]
-            return obj
-
     class DeleteChild(ModalFormMixin, OrgObjPermsMixin, SmartDeleteView):
         cancel_url = "@orgs.org_sub_orgs"
         success_url = "@orgs.org_sub_orgs"
@@ -1877,7 +1556,7 @@ class OrgCRUDL(SmartCRUDL):
         # invalid form login 'logs out' the user from the org and takes them to the org manage page
         def form_invalid(self, form):
             switch_to_org(self.request, None)
-            return HttpResponseRedirect(reverse("orgs.org_manage"))
+            return HttpResponseRedirect(reverse("staff.org_list"))
 
     class SubOrgs(SpaMixin, ContextMenuMixin, OrgPermsMixin, InferOrgMixin, SmartListView):
         title = _("Workspaces")
@@ -1991,7 +1670,7 @@ class OrgCRUDL(SmartCRUDL):
 
             if not org:
                 if user.is_staff:
-                    return HttpResponseRedirect(reverse("orgs.org_manage"))
+                    return HttpResponseRedirect(reverse("staff.org_list"))
 
                 return HttpResponseRedirect(reverse("orgs.org_choose"))
 
@@ -2024,7 +1703,7 @@ class OrgCRUDL(SmartCRUDL):
 
                 elif user_orgs.count() == 0:
                     if user.is_staff:
-                        return HttpResponseRedirect(reverse("orgs.org_manage"))
+                        return HttpResponseRedirect(reverse("staff.org_list"))
 
                     # for regular users, if there's no orgs, log them out with a message
                     messages.info(request, _("No workspaces for this account, please contact your administrator."))
@@ -2110,9 +1789,7 @@ class OrgCRUDL(SmartCRUDL):
             user = authenticate(username=user.username, password=self.form.cleaned_data["password"])
             login(self.request, user)
 
-            obj.add_user(user, self.invitation.role)
-
-            self.invitation.release()
+            self.invitation.accept(user)
 
     class JoinAccept(NoNavMixin, InvitationMixin, SmartUpdateView):
         """
@@ -2145,9 +1822,7 @@ class OrgCRUDL(SmartCRUDL):
             return None
 
         def save(self, obj):
-            obj.add_user(self.request.user, self.invitation.role)
-
-            self.invitation.release()
+            self.invitation.accept(self.request.user)
 
             switch_to_org(self.request, obj)