+++ temba/staff/urls.py
@@ -1,0 +1,4 @@
+from .views import OrgCRUDL, UserCRUDL
+
+urlpatterns = OrgCRUDL().as_urlpatterns()
+urlpatterns += UserCRUDL().as_urlpatterns()